Refactor micropkg and remove pipeline pull and pipeline package (#1328)
